+++ packages/camera/camera_android_camerax/test/android_camera_camerax_test.dart
@@ -215,8 +215,7 @@ void main() {
         name: 'cameraName',
         lensDirection: testLensDirection,
         sensorOrientation: testSensorOrientation);
-    const ResolutionPreset testResolutionPreset = ResolutionPreset.veryHigh;
-    const bool enableAudio = true;
+
     const int testSurfaceTextureId = 6;
 
     // Mock/Detached objects for (typically attached) objects created by
@@ -290,8 +289,16 @@ void main() {
     camera.processCameraProvider = mockProcessCameraProvider;
 
     expect(
-        await camera.createCamera(testCameraDescription, testResolutionPreset,
-            enableAudio: enableAudio),
+        await camera.createCameraWithSettings(
+          testCameraDescription,
+          const MediaSettings(
+            resolutionPreset: ResolutionPreset.low,
+            fps: 15,
+            videoBitrate: 200000,
+            audioBitrate: 32000,
+            enableAudio: true,
+          ),
+        ),
         equals(testSurfaceTextureId));
 
     // Verify permissions are requested and the camera starts listening for device orientation changes.
@@ -400,8 +407,15 @@ void main() {
         .thenAnswer((_) async => mockCameraControl);
     camera.processCameraProvider = mockProcessCameraProvider;
 
-    await camera.createCamera(testCameraDescription, testResolutionPreset,
-        enableAudio: enableAudio);
+    await camera.createCameraWithSettings(
+        testCameraDescription,
+        const MediaSettings(
+          resolutionPreset: testResolutionPreset,
+          fps: 15,
+          videoBitrate: 2000000,
+          audioBitrate: 64000,
+          enableAudio: enableAudio,
+        ));
 
     // Verify expected UseCases were bound.
     verify(camera.processCameraProvider!.bindToLifecycle(camera.cameraSelector!,
@@ -507,8 +521,11 @@ void main() {
 
     // Test non-null resolution presets.
     for (final ResolutionPreset resolutionPreset in ResolutionPreset.values) {
-      await camera.createCamera(testCameraDescription, resolutionPreset,
-          enableAudio: enableAudio);
+      await camera.createCamera(
+        testCameraDescription,
+        resolutionPreset,
+        enableAudio: enableAudio,
+      );
 
       Size? expectedBoundSize;
       ResolutionStrategy? expectedResolutionStrategy;
@@ -701,8 +718,6 @@ void main() {
         name: 'cameraName',
         lensDirection: testLensDirection,
         sensorOrientation: testSensorOrientation);
-    const ResolutionPreset testResolutionPreset = ResolutionPreset.veryHigh;
-    const bool enableAudio = true;
     const int resolutionWidth = 350;
     const int resolutionHeight = 750;
     final Camera mockCamera = MockCamera();
@@ -780,8 +795,16 @@ void main() {
     when(mockPreview.getResolutionInfo())
         .thenAnswer((_) async => testResolutionInfo);
 
-    await camera.createCamera(testCameraDescription, testResolutionPreset,
-        enableAudio: enableAudio);
+    await camera.createCameraWithSettings(
+      testCameraDescription,
+      const MediaSettings(
+        resolutionPreset: ResolutionPreset.medium,
+        fps: 15,
+        videoBitrate: 200000,
+        audioBitrate: 32000,
+        enableAudio: true,
+      ),
+    );
 
     // Start listening to camera events stream to verify the proper CameraInitializedEvent is sent.
     camera.cameraEventStreamController.stream.listen((CameraEvent event) {
@@ -870,7 +893,7 @@ void main() {
   });
 
   test(
-      'onDeviceOrientationChanged stream emits changes in device oreintation detected by system services',
+      'onDeviceOrientationChanged stream emits changes in device orientation detected by system services',
       () async {
     final AndroidCameraCameraX camera = AndroidCameraCameraX();
     final Stream<DeviceOrientationChangedEvent> eventStream =
@@ -1861,6 +1884,7 @@ void main() {
             as Analyzer;
 
     await capturedAnalyzer.analyze(mockImageProxy);
+
     final CameraImageData imageData = await imageDataCompleter.future;
 
     // Test Analyzer correctly process ImageProxy instances.