+++ packages/camera/camera_avfoundation/ios/Classes/CameraPlugin_Test.h
@@ -47,5 +47,4 @@
 /// @param result a thread safe flutter result wrapper object to report creation result.
 - (void)createCameraOnSessionQueueWithCreateMethodCall:(FlutterMethodCall *)createMethodCall
                                                 result:(FLTThreadSafeFlutterResult *)result;
-
 @end