+++ packages/camera/camera_windows/windows/record_handler.cpp
@@ -114,6 +114,22 @@ HRESULT BuildMediaTypeForAudioCapture(IMFMediaType** audio_record_media_type) {
   return hr;
 }
 
+// Helper function to set the frame rate on a video media type.
+inline HRESULT SetFrameRate(IMFMediaType* pType, UINT32 numerator,
+                            UINT32 denominator) {
+  return MFSetAttributeRatio(pType, MF_MT_FRAME_RATE, numerator, denominator);
+}
+
+// Helper function to set the video bitrate on a video media type.
+inline HRESULT SetVideoBitrate(IMFMediaType* pType, UINT32 bitrate) {
+  return pType->SetUINT32(MF_MT_AVG_BITRATE, bitrate);
+}
+
+// Helper function to set the audio bitrate on an audio media type.
+inline HRESULT SetAudioBitrate(IMFMediaType* pType, UINT32 bitrate) {
+  return pType->SetUINT32(MF_MT_AUDIO_AVG_BYTES_PER_SECOND, bitrate);
+}
+
 HRESULT RecordHandler::InitRecordSink(IMFCaptureEngine* capture_engine,
                                       IMFMediaType* base_media_type) {
   assert(!file_path_.empty());
@@ -160,6 +176,17 @@ HRESULT RecordHandler::InitRecordSink(IMFCaptureEngine* capture_engine,
     return hr;
   }
 
+  if (media_settings_.fps.has_value()) {
+    assert(media_settings_.fps.value() > 0);
+    SetFrameRate(video_record_media_type.Get(), media_settings_.fps.value(), 1);
+  }
+
+  if (media_settings_.video_bitrate.has_value()) {
+    assert(media_settings_.video_bitrate.value() > 0);
+    SetVideoBitrate(video_record_media_type.Get(),
+                    media_settings_.video_bitrate.value());
+  }
+
   DWORD video_record_sink_stream_index;
   hr = record_sink_->AddStream(
       (DWORD)MF_CAPTURE_ENGINE_PREFERRED_SOURCE_STREAM_FOR_VIDEO_RECORD,
@@ -168,13 +195,19 @@ HRESULT RecordHandler::InitRecordSink(IMFCaptureEngine* capture_engine,
     return hr;
   }
 
-  if (record_audio_) {
+  if (media_settings_.record_audio) {
     ComPtr<IMFMediaType> audio_record_media_type;
     HRESULT audio_capture_hr = S_OK;
     audio_capture_hr =
         BuildMediaTypeForAudioCapture(audio_record_media_type.GetAddressOf());
 
     if (SUCCEEDED(audio_capture_hr)) {
+      if (media_settings_.audio_bitrate.has_value()) {
+        assert(media_settings_.audio_bitrate.value() > 0);
+        SetAudioBitrate(audio_record_media_type.Get(),
+                        media_settings_.audio_bitrate.value());
+      }
+
       DWORD audio_record_sink_stream_index;
       hr = record_sink_->AddStream(
           (DWORD)MF_CAPTURE_ENGINE_PREFERRED_SOURCE_STREAM_FOR_AUDIO,