+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.9.15
+
+* Adds support to control video FPS and bitrate. See `CameraController.withSettings`.
+
 ## 0.9.14+1
 
 * Fixes bug where max resolution preset does not produce highest available resolution on iOS.