+++ packages/camera/camera_web/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 0.3.3
 
+* Adds support to control video FPS and bitrate. See `CameraController.withSettings`.
 * Updates minimum supported SDK version to Flutter 3.13/Dart 3.1.
 
 ## 0.3.2+4